aider: Implemented the `User` type according to the API documentation.
# Aider chat conversation:

USER: please also implemnt the user type according to the docs

ASSISTANT: To implement the `User` type according to the API documentation, we will need to create a new Go struct in the `api/types.go` file. This struct will include all the fields as described in the API documentation.

Here is the *SEARCH/REPLACE block* for adding the `User` struct to the `api/types.go` file:

api/types.go
```go
<<<<<<< SEARCH
=======
// User represents a user in the Vikunja API.
type User struct {
	ID               int    `json:"id"`
	Username         string `json:"username"`
	Name             string `json:"name"`
	DefaultProjectID int    `json:"default_project_id"`
}
>>>>>>> REPLACE
```

This `User` struct includes the fields `ID`, `Username`, `Name`, and `DefaultProjectID`, which are the basic fields for a user according to the provided API documentation. If there are additional fields or specific types that need to be used, please provide that information, and I can adjust the struct accordingly.