aider: Updated edit command to use `EditStringInEditor` function for editing the title and description fields of a task.
# Aider chat conversation:

USER: The utility function should now be present. Now, in the edit task command, please use this `EditStringInEditor`function to edit the description field and the title field, supplying the previous value of thse fields as initial content for the edotor

ASSISTANT: To implement the requested changes, we need to modify the `cmd/task.go` file to use the `EditStringInEditor` function for editing the title and description fields of a task. Here are the steps and the corresponding *SEARCH/REPLACE* blocks:

1. Update the `editCmd` command to use `EditStringInEditor` for editing the title and description.

cmd/task.go
```go
<<<<<<< SEARCH
case "Title":
	prompt := &survey.Input{Message: "Enter new title:"}
	survey.AskOne(prompt, &task.Title)
case "Description":
	prompt := &survey.Input{Message: "Enter new description:"}
	survey.AskOne(prompt, &task.Description)
=======
case "Title":
	editedTitle, err := EditStringInEditor(task.Title)
	if err != nil {
		fmt.Println("Error editing title:", err)
		continue
	}
	task.Title = editedTitle
case "Description":
	editedDescription, err := EditStringInEditor(task.Description)
	if err != nil {
		fmt.Println("Error editing description:", err)
		continue
	}
	task.Description = editedDescription
>>>>>>> REPLACE
```

This change will replace the existing prompts for the title and description with calls to `EditStringInEditor`, passing the current title and description as the initial content for the editor. If an error occurs during the editing process, it will print an error message and continue the loop, allowing the user to try editing again or choose another option.